split up workflow renderer into 2 files
the size of at least the output renderers will grow significantly in the
future
